Use cssparser's new_with_line_number_offset
cssparser provides a way to set the initial line number on a
ParserInput.  This patch changes servo to use this facility, rather than
reimplement the same functionality itself.

+++ components/script/dom/htmlstyleelement.rs
@@ -98,7 +98,7 @@ impl HTMLStyleElement {
                                          shared_lock, Some(&loader),
                                          win.css_error_reporter(),
                                          doc.quirks_mode(),
-                                         self.line_number);
+                                         self.line_number as u32);
 
         let sheet = Arc::new(sheet);
